+++ core/autoencoder.cpp
@@ -84,14 +84,7 @@ namespace oidn {
 
     device->executeTask([&]()
     {
-      if (hdr)
-      {
-        const float exposure = autoexposure(color);
-        //printf("exposure = %f\n", exposure);
-        std::static_pointer_cast<HDRTransferFunction>(transferFunc)->setExposure(exposure);
-      }
-
-      net->execute();
+      net->execute(progressFunc, progressUserPtr);
     });
   }
 
@@ -232,14 +225,18 @@ namespace oidn {
     std::shared_ptr<Node> inputReorder;
     if (srgb)
     {
-      transferFunc = std::make_shared<LinearTransferFunction>();
+      transferFunc = std::make_shared<LDRLinearTransferFunction>();
       inputReorder = net->addInputReorder(color, albedo, normal,
-                                          std::static_pointer_cast<LinearTransferFunction>(transferFunc),
+                                          std::static_pointer_cast<LDRLinearTransferFunction>(transferFunc),
                                           spatialPad, inputReorderDst);
     }
     else if (hdr)
     {
       transferFunc = std::make_shared<HDRTransferFunction>();
+
+      net->addAutoexposure(color,
+                           std::static_pointer_cast<HDRTransferFunction>(transferFunc));
+
       inputReorder = net->addInputReorder(color, albedo, normal,
                                           std::static_pointer_cast<HDRTransferFunction>(transferFunc),
                                           spatialPad, inputReorderDst);
@@ -348,7 +345,7 @@ namespace oidn {
 
     // Output reorder
     if (srgb)
-      net->addOutputReorder(conv11->getDst(), std::static_pointer_cast<LinearTransferFunction>(transferFunc), output);
+      net->addOutputReorder(conv11->getDst(), std::static_pointer_cast<LDRLinearTransferFunction>(transferFunc), output);
     else if (hdr)
       net->addOutputReorder(conv11->getDst(), std::static_pointer_cast<HDRTransferFunction>(transferFunc), output);
     else