+++ core/filter.cpp
@@ -14,37 +14,14 @@
 // limitations under the License.                                           //
 // ======================================================================== //
 
-#pragma once
-
-#include "common.h"
-#include "device.h"
-#include "image.h"
+#include "filter.h"
 
 namespace oidn {
 
-  class Filter : public RefCount
+  void Filter::setProgressMonitorFunction(ProgressMonitorFunction func, void* userPtr)
   {
-  protected:
-    Ref<Device> device;
-
-    ProgressMonitorFunction progressFunc = nullptr;
-    void* progressUserPtr = nullptr;
-
-    bool dirty = true;
-
-  public:
-    explicit Filter(const Ref<Device>& device) : device(device) {}
-
-    virtual void setImage(const std::string& name, const Image& data) = 0;
-    virtual void set1i(const std::string& name, int value) = 0;
-    virtual int get1i(const std::string& name) = 0;
-
-    void setProgressMonitorFunction(ProgressMonitorFunction func, void* userPtr);
-
-    virtual void commit() = 0;
-    virtual void execute() = 0;
-
-    Device* getDevice() { return device.get(); }
-  };
+    progressFunc = func;
+    progressUserPtr = userPtr;
+  }
 
 } // namespace oidn