+++ core/tone_mapping.h
@@ -17,6 +17,7 @@
 #pragma once
 
 #include "image.h"
+#include "node.h"
 
 namespace oidn {
 
@@ -35,11 +36,19 @@ namespace oidn {
     virtual float inverse(float x) const = 0;
   };
 
-  class LinearTransferFunction : public TransferFunction
+  // LDR transfer function: linear
+  class LDRLinearTransferFunction : public TransferFunction
   {
   public:
-    __forceinline float forward(float y) const override { return y; }
-    __forceinline float inverse(float x) const override { return x; }
+    __forceinline float forward(float y) const override
+    {
+      return min(y, 1.f);
+    }
+
+    __forceinline float inverse(float x) const override
+    {
+      return min(x, 1.f);
+    }
   };
 
   // LDR transfer function: sRGB curve
@@ -48,12 +57,12 @@ namespace oidn {
   public:
     __forceinline float forward(float y) const override
     {
-      return pow(y, 1.f/2.2f);
+      return pow(min(y, 1.f), 1.f/2.2f);
     }
 
     __forceinline float inverse(float x) const override
     {
-      return pow(x, 2.2f);
+      return min(pow(x, 2.2f), 1.f);
     }
   };
 
@@ -128,6 +137,29 @@ namespace oidn {
     }
   };
 
-  float autoexposure(const Image& color);
+  // Autoexposure node
+  class AutoexposureNode : public Node
+  {
+  private:
+    Image color;
+    std::shared_ptr<HDRTransferFunction> transferFunc;
+
+  public:
+    AutoexposureNode(const Image& color,
+                     const std::shared_ptr<HDRTransferFunction>& transferFunc)
+      : color(color),
+        transferFunc(transferFunc)
+    {}
+
+    void execute(stream& sm) override
+    {
+      const float exposure = autoexposure(color);
+      //printf("exposure = %f\n", exposure);
+      transferFunc->setExposure(exposure);
+    }
+
+  private:
+    static float autoexposure(const Image& color);
+  };
 
 } // namespace oidn